+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated jsonoid-discovery dependency to v0.7.3
+- Allow undefined properties to be required
 
 ### Fixed
 - Fix assembly task